feat(components): [virtual-table] compsables (#7341)
- Split `use-table` into separate files for better readability
